Finished levelselction for now
#531
time: 1
state: Done
comment: After a lot of experiments, it was decided to use cards instead of world objects for levelselection. Using worldobjects tended to cause issues with the bahavior of the UI elements as well as with the positioning of the worlds themselves. There might be coming a rework, in which we reconfigure the levelselection screen to fit the illustration of mini-worlds, but for now, with regards on the remaining time, this is it.