Experiments on Lvl3
Specific players have to activate a specific plate "hidden" on the map. When triggered, one of the blocks rises, so that they can reach the zone and therfore the targetzone. In addition to that, the plates have to be activated in a specific order. If the wrong order happened to be, the blocks are lowered again and the players have to start from the beginning to activate the blocks. Unfortunately, once recognized a wrong order has been entered, the logic does not work anymore. IndexOutOfBound is also present when activating more triggers than expected - this will simply be avoided by setting the zone / trigger deactive after triggered.