Working Settingsscreen
#554
time:2
comment:user is able to switch between window and fullscreen mode. The mastervolume can be changed and works thorugh all different scenes. Graphics can be switched between low, middle, high and ultra. Possible enhancement could be a dropdown for resolutions. Known issues, mastervolume is not automatically updated when returning back to the mainmenu. furthermore the soundfiles do not use 100 per cent of the possible output, so that music stops playing although the slider is not completely down