Working Collectible for Devil
#586
comment: devil is now able to collect ability of angel. demo is seen in lvl1. prefab mesh has to be changed and therefore the trigger boundaries have to be adjusted. For now, players use then the same key for jumping (space), so that they have to jump together and in order to do this properly - work together as a team in front of the pc.