Sprint_11: AI Approach
#557
time:2
comment: Adjusted NavMesh to work more accurate. Therefore splitted the different blogs and allowed the ai to move between them using a connector. Furthermore applied the Logic to a placeholder, so that the devil is able to strike the AI down. In addition to that there are first steps on how the AI should behave and follow the player - unfortunately this approach throws error by now and needs more adjustments. This will be achieved within the next submit.