Sprint_11: Working AI
#557
time:1
state:Done
comment:AI does now work properly. When in Range, it chases the player who has been in range first. When then closer to it, health of player is taken. Stopping distance does now work properly as well as moving over the brigdes. When stopped, AI turn correctly to look towards the player. Only adjustment needed is the changing of the placeholder. Here the real soul model should be applied, but this is done within a few seconds, so nothing to worry about.

#555
state:Done